+++ HelloReact/ios/HelloReact/AppDelegate.m
@@ -1,12 +1,3 @@
-/**
- * Copyright (c) 2015-present, Facebook, Inc.
- * All rights reserved.
- *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree. An additional grant
- * of patent rights can be found in the PATENTS file in the same directory.
- */
-
 #import "AppDelegate.h"
 
 #import <React/RCTBundleURLProvider.h>
@@ -33,11 +24,11 @@
   rootViewController.view = rootView;
   self.window.rootViewController = rootViewController;
   [self.window makeKeyAndVisible];
-  
+
   [Appboy startWithApiKey:@"e5cbdeaf-00b9-4a22-8587-815e6dbcf520"
             inApplication:application
         withLaunchOptions:launchOptions];
-  
+
   return YES;
 }